Split native motion DTO module
